requirements.txt
<<<<<<< HEAD
# pip install -r requirements.txt
OpenAI
llama-index
google-api-python-client
youtube-transcript-api
llama_index
langchain
sentence_transformers
=======
aiohttp==3.9.1
aiosignal==1.3.1
aiostream==0.5.2
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.3
datasets==2.15.0
Deprecated==1.2.14
dill==0.3.7
distro==1.8.0
dnspython==2.4.2
exceptiongroup==1.2.0
filelock==3.13.1
frozenlist==1.4.0
fsspec==2023.10.0
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.19.4
idna==3.6
joblib==1.3.2
llama-index==0.9.8.post1
loguru==0.7.2
marshmallow==3.20.1
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nest-asyncio==1.5.8
nltk==3.8.1
numpy==1.26.2
openai==1.3.5
packaging==23.2
pandas==2.1.3
pinecone-client==2.2.4
pyarrow==14.0.1
pyarrow-hotfix==0.6
pydantic==2.5.2
pydantic_core==2.14.5
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
tenacity==8.2.3
tiktoken==0.5.1
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
win32-setctime==1.1.0
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.3
>>>>>>> 7fb3e9107b1659f7b62565430e46184a9dfed58f
